+++ instrumentation/rabbitmq-2.7/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/rabbitmq/v2_7/RabbitReceiveAttributesGetter.java
@@ -13,6 +13,8 @@ import static java.util.Collections.singletonList;
 
 import com.rabbitmq.client.GetResponse;
 import io.opentelemetry.instrumentation.api.incubator.semconv.messaging.MessagingAttributesGetter;
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import javax.annotation.Nullable;
@@ -112,5 +114,15 @@ final class RabbitReceiveAttributesGetter
       return emptyList();
     }
     return singletonList(value.toString());
+  }
+
+  @Override
+  public Collection<String> getMessageHeaderNames(ReceiveRequest request) {
+    GetResponse response = request.getResponse();
+    if (response == null) {
+      return emptyList();
+    }
+    Map<String, Object> headers = response.getProps().getHeaders();
+    return headers == null ? emptyList() : new ArrayList<>(headers.keySet());
   }
 }